refactor(challenge): ChallengeRoutineBatchUpdateIntegrationTest.java 분리
